pyproject.toml
[project]
name = "filesharing"
version = "0.1.0"
description = "File sharing system for discord users"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
